requirements.txt
pillow
geopy
openai
dashscope
bs4
fastmcp
chainlit